Made a lot of changes to results_retrieval but now I'm getting a serious error when I try to launch Eclipse

+++ eclipseplugin/src/org/continuousassurance/swamp/eclipse/ResultsRetriever.java
@@ -7,12 +7,10 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
-
 import org.continuousassurance.swamp.eclipse.exceptions.ResultsRetrievalException;
 import org.continuousassurance.swamp.eclipse.exceptions.UserNotLoggedInException;
 import org.continuousassurance.swamp.eclipse.ui.StatusView;
 import org.continuousassurance.swamp.eclipse.ui.SwampPerspective;
-import org.eclipse.core.runtime.Status;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.PlatformUI;
@@ -21,6 +19,8 @@ import edu.uiuc.ncsa.swamp.api.AssessmentRecord;
 import edu.wisc.cs.swamp.SwampApiWrapper;
 
 public class ResultsRetriever {
+	
+	private static final String USER_NOT_LOGGED_IN_STATUS = "User not logged in";
 
 	public static void retrieveResults() throws UserNotLoggedInException, ResultsRetrievalException {
 		// (1) If user's not logged in, quit out immediately
@@ -28,28 +28,42 @@ public class ResultsRetriever {
 			throw new UserNotLoggedInException();
 		}
 	
-		// (2) Read through unfinished file
+		SwampApiWrapper api = null;
+		try {
+			api = new SwampApiWrapper(SwampApiWrapper.HostType.CUSTOM, Activator.getLastHostname());
+			if (!api.restoreSession()) {
+				throw new UserNotLoggedInException();
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw new UserNotLoggedInException();
+		}
+
+		List<String> statuses = new ArrayList<>();
+		addUnfinishedFileStatuses(statuses, api);
+		addFinishedFileStatuses(statuses);
+		
+		System.out.println("Updating status view");
+		updateStatusView(statuses);
+		return;
+	}
+	
+	public static void retrieveResultsNotLoggedIn() {
+		
+	}
+	
+	public static void addUnfinishedFileStatuses(List<String> statuses, SwampApiWrapper api) throws ResultsRetrievalException {
 		File oldFile = new File(Activator.getUnfinishedAssessmentsPath());
+		if (!oldFile.exists()) {
+			return;
+		}
 		Scanner sc = null;
 		try {
 			sc = new Scanner(oldFile);
 		} catch (FileNotFoundException e) {
 			return;
 		}
-
-		SwampApiWrapper api = null;
-		try {
-			api = new SwampApiWrapper(SwampApiWrapper.HostType.CUSTOM, Activator.getLastHostname());
-			if (!api.restoreSession()) {
-				sc.close();
-				throw new UserNotLoggedInException();
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-			sc.close();
-			throw new UserNotLoggedInException();
-		}
-
+		
 		File tmp = new File(Activator.getUnfinishedAssessmentsPath() + ".tmp");
 		if (tmp.exists()) {
 			tmp.delete();
@@ -65,7 +79,6 @@ public class ResultsRetriever {
 		
 	
 		System.out.println("Now to parse stuff from unfinished file");
-		List<String> statuses = new ArrayList<>();
 		while (sc.hasNextLine()) {
 			String assessmentDetails = sc.nextLine();
 			System.out.println("New line: " + assessmentDetails);
@@ -88,16 +101,13 @@ public class ResultsRetriever {
 		oldFile.delete();
 		
 		tmp.renameTo(oldFile);
-		
-		addFinishedFileStatuses(statuses);
-		
-		System.out.println("Updating status view");
-		updateStatusView(statuses);
-		return;
 	}
 	
 	private static void addFinishedFileStatuses(List<String> statuses) {
 		File f = new File(Activator.getFinishedAssessmentsPath());
+		if (!f.exists()) {
+			return;
+		}
 		Scanner sc = null;
 		try {
 			sc = new Scanner(f);
@@ -132,24 +142,43 @@ public class ResultsRetriever {
 			System.out.println("No statuses");
 		}
 	}
+	
+	private static void writeToFinishedFile(String newDetailInfo) throws IOException {
+		File f = new File(Activator.getFinishedAssessmentsPath());
+		if (!f.exists()) {
+			f.createNewFile();
+		}
+		FileWriter finishedWriter = new FileWriter(f, true);
+		finishedWriter.write(newDetailInfo);
+		finishedWriter.close();
+	}
 
 	private static String updateStatus(FileWriter unfinishedWriter, SwampApiWrapper api, String prjUUID, String assessUUID, 
 			String serializedAssessmentDetails) {
 		System.out.println("Querying for assessment record with project UUID " + prjUUID + " and assessment UUID " + assessUUID);
+		if (api == null) {
+			return AssessmentDetails.updateStatus(serializedAssessmentDetails, USER_NOT_LOGGED_IN_STATUS);
+		}
 		AssessmentRecord rec = api.getAssessmentRecord(prjUUID, assessUUID);
 		String status = rec.getStatus();
 		System.out.println("Status: " + status);
 		String newDetailInfo = AssessmentDetails.updateStatus(serializedAssessmentDetails, status);
-		FileWriter finishedWriter = null;
 		try {
-			if ("Complete".equals(status)) { // TODO: Fix this when Vamshi modifies the AssessmentRecord.java API. This is NOT how we should be doing it
-				File f = new File(Activator.getFinishedAssessmentsPath());
-				if (!f.exists()) {
-					f.createNewFile();
+			if ("Finished".equals(status)) { // TODO: Fix this when Vamshi modifies the AssessmentRecord.java API. This is NOT how we should be doing it
+				System.out.println("Finished with no errors!");
+				newDetailInfo = AssessmentDetails.addBugCount(serializedAssessmentDetails, Integer.toString(rec.getWeaknessCount()));
+				String filepath = AssessmentDetails.getFilepath(serializedAssessmentDetails);
+				File f = new File(filepath);
+				if (f.exists()) {
+					f.delete();
 				}
-				finishedWriter = new FileWriter(f, true);
-				finishedWriter.write(newDetailInfo);
-				finishedWriter.close();
+				api.getAssessmentResults(prjUUID, assessUUID, filepath);
+				writeToFinishedFile(newDetailInfo);
+				return null;
+			}
+			else if ("Finished with Errors".equals(status)) { // I've left these hardcoded and without a var as a reminder that this needs to be fixed ASAP. This will break as soon as they change the status labels
+				System.out.println("Finished with errors!");
+				writeToFinishedFile(newDetailInfo);
 				return null;
 			}
 			else {
